adds forecast POJO with cunstructor layout
